Move agent init to lib
